Update docs to match current implementation.
The prog ref docs use javascrip and css to dynamically remove features of
pysvn that depend on particular versions of SVN.

Update generator script to only extract documentation depend on the version
of svn that is being built against. Thus the docs builtin to the pysvn extension
will correct docuement the API.

update setup.py and the make files to pass the svn include dir to the generator.


git-svn-id: https://svn.code.sf.net/p/pysvn/code/trunk/pysvn/Extension@528 8a6430f3-a0d1-4a6a-aeff-73e05446c159