patch 1/3 from Bert Wesarg
Writing a pre-commit hook with pysvn is a pleasure. Unfortunatly you have to
use the normal client API for post-commit hooks. Subversion itself provides
the same API for pre- and post-commit hooks with its svnlook command. Having
the same API for both hooks just feels right and also makes it easy to test
your pre-commit hook on a revision. This patch implements this by providing
an optional parameter to pysvn.Transaction() to let it work on a revision
instead of a transaction.

The test uses the same (now renamed) commit hook code for pre- and post-commit.


git-svn-id: https://svn.code.sf.net/p/pysvn/code/trunk/pysvn/Extension@1331 8a6430f3-a0d1-4a6a-aeff-73e05446c159