Patch from Robie Basak for issue #184.
On the current development release of Ubuntu, we're seeing a build failure
because pyconfig.h is in /usr/include/x86_64-linux-gnu/python2.7 but Python.h is
in /usr/include/python2.7. This is to support multiarch. However, the pysvn
build system assumes that both arch-specific and arch-independent includes are
in the same directory, contrary to the specification of
distutils.sysconfig.get_python_inc.


git-svn-id: https://svn.code.sf.net/p/pysvn/code/trunk/pysvn/Extension@1503 8a6430f3-a0d1-4a6a-aeff-73e05446c159